+++ brigade/plugins/tasks/commands/__init__.py
@@ -1,0 +1,6 @@
+from remote_command import remote_command
+
+
+__all__ = (
+    "remote_command",
+)
