+++ brigade/plugins/tasks/data/__init__.py
@@ -1,0 +1,8 @@
+from .load_json import load_json
+from .load_yaml import load_yaml
+
+
+__all__ = (
+    "load_json",
+    "load_yaml",
+)
