+++ brigade/plugins/tasks/networking/__init__.py
@@ -1,0 +1,7 @@
+from .napalm_configure import napalm_configure
+from .napalm_get_facts import napalm_get_facts
+
+__all__ = (
+    "napalm_configure",
+    "napalm_get_facts",
+)
