+++ brigade/plugins/tasks/networking/napalm_configure.py
