+++ brigade/plugins/tasks/networking/napalm_get_facts.py
