+++ brigade/plugins/tasks/text/__init__.py
@@ -1,0 +1,7 @@
+from .template_file import template_file
+from .template_string import template_string
+
+__all__ = (
+    "template_file",
+    "template_string",
+)
