+++ docs/tasks/data.rst
@@ -1,0 +1,6 @@
+Data
+====
+
+.. automodule:: brigade.plugins.tasks.data
+   :members:
+   :undoc-members:
