+++ docs/tasks/networking.rst
@@ -1,0 +1,6 @@
+Networking
+==========
+
+.. automodule:: brigade.plugins.tasks.networking
+   :members:
+   :undoc-members:
