+++ docs/tasks/text.rst
@@ -1,0 +1,6 @@
+Text
+====
+
+.. automodule:: brigade.plugins.tasks.text
+   :members:
+   :undoc-members:
